Goa/*: change unsigned long ip to unsigned int ip per sezero advice for LP64.

+++ Goa/CEngine/gserver.c
@@ -35,7 +35,7 @@ void ServerFree(void *elem)
 	free(server);
 }
 
-GServer ServerNew(unsigned long ip, unsigned short port)
+GServer ServerNew(unsigned int ip, unsigned short port)
 {
 	GServer server;
